Refactor codebase into proper packages
