+++ src/main/java/data/tasks.txt
@@ -1,0 +1,2 @@
+D|0|return book|2/12/2019 1800
+E|0|team meeting|15/3/2020 1400|15/3/2020 1600
